Application2/backend/requirements.txt
fastapi
uvicorn
pymilvus
langchain-huggingface
langchain-milvus
langchain-community
PyPDF2
python-dotenv
PyJWT